+++ CHANGELOG.md
@@ -8,6 +8,12 @@ You must updated your config when updating from 0.1.13 to 1.0.0 (or later versio
 
 # Bug Fixes and Improvements
 
+## 1.0.3-beta.6 (2021-03-23)
+### Changes
+* Added support of device syncName and name, see Readme for details
+* Improved handling of device names when syncing as device accepts names from 3 to 14 characters only
+
+
 ## 1.0.2 (2021-03-22)
 ### Changes
 * Improved handling when no supported devices found